requirements.txt
django
openai
gunicorn
python-dotenv
assemblyai